+++ main.py
@@ -44,6 +44,16 @@ class Main(KytosNApp):
     def execute(self):
         """ Topology updates are executed through events. """
 
+    @listen_to('kytos/topology.switch.disabled')
+    def on_switch_disabled(self, event):
+        """Remove switch from self.switches"""
+        self.handle_switch_disabled(event.content['dpid'])
+
+    @listen_to('kytos/topology.link.disabled')
+    def on_link_disabled(self, event):
+        """Remove link from self.switches neighbors"""
+        self.handle_link_disabled(event.content['link'])
+
     @listen_to('kytos/topology.updated')
     def topology_updated(self, event):
         """Update colors on topology update."""
@@ -52,6 +62,7 @@ class Main(KytosNApp):
             [link.as_dict() for link in topology.links.values()]
         )
 
+    # pylint: disable=too-many-branches
     def update_colors(self, links):
         """ Color each switch, with the color based on the switch's DPID.
             After that, if not yet installed, installs, for each switch, flows
@@ -60,6 +71,10 @@ class Main(KytosNApp):
         """
         with self._switches_lock:
             for switch in self.controller.switches.copy().values():
+                if not switch.is_enabled():
+                    if switch.dpid in self.switches:
+                        self.switches[switch.dpid]['neighbors'] = set()
+                    continue
                 if switch.dpid not in self.switches:
                     color = int(switch.dpid.replace(':', '')[4:], 16)
                     self.switches[switch.dpid] = {'color': color,
@@ -69,6 +84,8 @@ class Main(KytosNApp):
                     self.switches[switch.dpid]['neighbors'] = set()
 
             for link in links:
+                if link.get('enabled') is not True:
+                    continue
                 source = link['endpoint_a']['switch']
                 target = link['endpoint_b']['switch']
                 if source != target:
@@ -82,11 +99,11 @@ class Main(KytosNApp):
         with self._switches_lock:
             for dpid, switch_dict in self.switches.items():
                 switch = self.controller.get_switch_by_dpid(dpid)
+                if switch.status != EntityStatus.UP:
+                    continue
                 if switch.ofp_version == '0x04':
                     controller_port = PortNo.OFPP_CONTROLLER
                 else:
-                    continue
-                if switch.status != EntityStatus.UP:
                     continue
                 for neighbor in switch_dict['neighbors']:
                     if neighbor not in switch_dict['flows']:
@@ -110,6 +127,49 @@ class Main(KytosNApp):
                         dpid_flows[dpid].append(flow_dict)
 
         self._send_flow_mods(dpid_flows)
+
+    def handle_link_disabled(self, link):
+        """Handle link deletion. Deletes only flows from the proper switches.
+         The field 'neighbors' is managed by update_colors method."""
+        switch_a_id = link.endpoint_a.switch.dpid
+        switch_b_id = link.endpoint_b.switch.dpid
+
+        if switch_b_id == switch_a_id:
+            return
+
+        with self._switches_lock:
+            flow_mods = defaultdict(list)
+            flow = self.switches[switch_a_id]['flows'][switch_b_id]
+            flow_mods[switch_a_id].append({
+                "table_id": flow['table_id'],
+                "match": flow['match']
+            })
+            flow = self.switches[switch_b_id]['flows'][switch_a_id]
+            flow_mods[switch_b_id].append({
+                "table_id": flow['table_id'],
+                "match": flow['match']
+            })
+            self.switches[switch_a_id]['flows'].pop(switch_b_id)
+            self.switches[switch_b_id]['flows'].pop(switch_a_id)
+        self._remove_flow_mods(flow_mods)
+
+    def handle_switch_disabled(self, dpid):
+        """Handle switch deletion. Links are expected to be disabled first
+         therefore the deleted inner dictionary is expected to be empty with
+         no flows and neighbors."""
+        with self._switches_lock:
+            try:
+                sw_dct = self.switches[dpid]
+                if sw_dct['flows'] or sw_dct['neighbors']:
+                    log.error(f"There was an error cleanning up {dpid}. "
+                              "The fields 'flows' and 'neighbors' should be"
+                              " empty.")
+                    return
+            except KeyError as err:
+                log.error(f"Error while handling disabled switch: "
+                          f"Switch {err} not found.")
+                return
+            self.switches.pop(dpid, None)
 
     def shutdown(self):
         """This method is executed when your napp is unloaded.
@@ -165,6 +225,18 @@ class Main(KytosNApp):
                 log.error(f'Flow manager returned an error inserting '
                           f'flows {flows}. Status code {res.status_code} '
                           f'on dpid {dpid}')
+
+    def _remove_flow_mods(self, flows: dict, force: bool = True) -> None:
+        """Remove FlowMods"""
+        for dpid, mod_flows in flows.items():
+            name = "kytos.flow_manager.flows.delete"
+            content = {
+                'dpid': dpid,
+                'flow_dict': {'flows': mod_flows},
+                'force': force,
+            }
+            event = KytosEvent(name=name, content=content)
+            self.controller.buffers.app.put(event)
 
     @rest('colors')
     def rest_colors(self, _request: Request) -> JSONResponse: